CreateSafePass Almost Almost Finished
Safe (not pwned) passes are created. Bugs:
If program cannot generate a safe pass even though the option is checked (e.g. ##### is not safe, but if user wants that pass and that length it is impossible)....
Situation by which 6 symbols are requested and length of pass is 5 -> impossible to use all requested chars